Fixed lifecycle issue with TxCommitProcess
Issue resulted in a master instance being healthy but unable to commit
any transactions because it's TransactionCommitProcess was not properly
switched on.

Commit process is switched on from a CommitProcessSwitcher which
registers itself as a listener in HighAvailabilityModeSwitcher. Latter
class notifies all registered listeners during slave and master
switches.

Whole clustering infrastructure started up and begin communication with
other machines concurrently with listener registration. This resulted in
a race between listener registration and notification.

This commit puts all ModeSwitcher listeners in a separate LifeSupport
instance which is started before clustering and paxos. Such startup
sequence ensures that all listeners are registered before clustering
communication starts.

Co-authored-by: @MishaDemianenko