Update lsof parsing for newer and older versions.
This change handles multiple pid numbers which can be attached to the
same port, which was not accounted for in the previous version of this
script. This change now handles the older versions of lsof and the
newer versions. The reason things broke was because the newer version
did not produce the expected output.