Fallback to TransactionIdStore.BASE_TX_ID in OnDiskLastTxIdGetter when neo stores are closed
OnDiskLastTxIdGetter is used for credentials creation in cluster election process.
It means that we should be ready to handle cases when neo stores are stopped: for example while we are copying store from master.
For those cases we will fallback to BASE_TX_ID as current latest commited transaction value.