Compute correct sampling of non-unique indexes
Sampling of non-unique indexes was incorrect since the code was
expecting to see multiple terms with the same values in the lucene
store.  This is actually incorrect since we don't duplicate value
entries in lucene, but we simply use them to index the node ids.

This commit will fix the sampling by looking at the frequency of the
terms containing the values.  Such frequency will tell us how many ids
are indexed by the same value.