Prevent lost updates on concurrent set property
Running multiple queries in parallel that writes a property with a value
that is based on reading the same property risks having an unpredictable result,
e.g. increment  "MATCH (n) SET n.prop = n.prop + 1".
We now make SET behave atomically by taking an exclusive lock on the entity
that we are going to write to before reading the value.
This decision is taken at runtime in the set property pipes.