Fix Cluster Uri regression
This was previously fixed in PR #5869, but that change broke
tyre-kicking somehow.

This is the minimum fix, which simply reverts part of
5f7bf49d546e8b91aafa11e3df8938a42d305bf9

This change has been verified to not break tyre-kicking, as well
as fixing the root cause (we shouldn't do a reverseDNS lookup of
the hostname of the local machine).