Make `Settings` utility class internal
Adding new configuration options to Neo4j is not a public API. This
commit deprecates the Settings public utility class in favor of an internal
equivalent class.

The old class does not inherit from the new one, since the intent of this
change (beyond cleaning up surface API) is to allow refactoring the
now-internal Settings class.