Introduce grouped configuration.
This allows configuration options like:

    dbms.books.0.name=Lord of the Rings
    dbms.books.0.author=Unknown
    dbms.books.1.name=Dr Who
    dbms.books.1.author=Spielberg, Steven

This commit does not contain the user-facing java API part of this, but it
will simply be a way to help build these setting strings, like:

    setConfig( books(0, book_name ), "Lordi" )

Eg. it ends up existing only once we use this API, which we will do for
Bolt in upcoming commits.