Make default dbms dir relative to db location.
For our production artifacts, we explicitly put the files that go under
`dbms` dir at their appropriate locations depending on platform.

However, for tests, there is a default directory (`data/dbms` and
`neo4j-home/dbms`) which is both obnoxious and inconsistent between
settings.

This introduces a base directory, relative to where the database directory
is. It does not change any production defaults, since those are specified
in the neo4j-server.properties file, but it means our tests and the
neo4j-harness module stops putting `neo4j-home` all over the place.