Fix flaky TestApps#canTerminateAnActiveCommand()
Test checks that it is possible to terminate running shell command by id.
It starts a lengthy Cypher query and separate thread to terminate this query.
Coordination between these two things is only done via Thread#sleep() which
is non-deterministic. This made test fail randomly.

This commit makes this test deterministic. Cypher query is not super long
running and command termination thread awaits for a single transaction to
be started (to avoid missing of termination signals).