Stricter constructor for ClassGuardedAdversary
ClassGuardedAdversary is an Adversary that does fault injection only when
specified classes are present in the stacktrace. It accepts vararg of class
names as a constructor parameter. This is a bit problematic because class
names could be given as strings and get stale. This happened in
PartialTransactionFailureIT which referenced a non-existing class from
the 'org.neo4j.kernel.impl.nioneo.xa' package.

This commit changes the constructor of ClassGuardedAdversary to take vararg
of Class objects.