Cleaned up store flushing and counts store rotation on shutdown
Lifecycle of all stores is basically managed by NeoStoreDataSource. It's
shutdown causes flush of all stores and rotation of the counts store.
Shutdown sequence heavily depends on the database state (kept by
DatabaseHealth class).

Currently during shutdown NeoStoreDataSource tries to flush stores two times.
First flush is using StoreFlusher directly and not guarded by the database
health check. Second flush is done by the checkpointer and is guarded by health
check. This causes counts store to be rotated two times as well.

As a last step in NeoStoreDataSource shutdown NeoStores are closed. This
triggers another rotation of the counts store. This second rotation seems to be
needed because previously there was a possibility of index sampler to make
changes to the counts store while NeoStoreDataSource was shutting down.
This is not necessary any more because we wait for index sampling to stop on
before the data source shutdown. Relevant commit is here #4831.

This commit removes the first store flush that is not guarded by the database
health check. It also removes rotation of the counts store from NeoStores.
With such change stores are flushed only once and counts store is rotated only
once on a healthy shutdown. On a panic shutdown stores are not flushed and
counts store is not rotated. In this case we rely on the recovery to fix all
possible pending/incomplete changes.