Fix deadlock in timeout service + minor cleanups.
The deadlock could occur when the timeout cancel operation was called
from a client inside the same monitor that the timeout expiry would be
handled under. So there was the possibility of a circular wait graph.

Both the timeout expiry and the cancelling operation were locking
the internal collection of timeouts during processing, which is unnecessary.
The deadlock loop is broken by not locking the timeouts collection when
processing the triggered timeouts.

This commit also cleans up some logging and dependency injection.