StorageEngine API lives in its own component
initially started merely as an interface inside the kernel component,
but now pulled out into neo4j-storage-engine-api. The contents of
this component is the StorageEngine interface itself and whatever that
API depends on, either as-is or subsets of those interfaces extracted from
kernel as well:
- StoreReadLayer which sits as read access to the underlying storage
  right under the kernel API.
- [kernel] exceptions which are checked exceptions, part of kernel API.
  StoreReadLayer uses the same exceptions so that package moved into
  this component as well. In the end this package may warrant its own
  component as a whole.
- Subset of interface for locking nodes and relationships. This due to
  the way the current storage engine implementation is design where some
  entities will have to be locked during the time of building record changes
  as part of commit.
- Schema interfaces that StoreReadLayer provides, such as IndexDescriptor,
  IndexReader, LabelScanReader etc.
- Reading of TransactionState which is one of the responsibilities that
  a StorageEngine has; to create commands from state representing changes
  made in a transaction.

Ultimate purpose of breaking this API out into its own component is twofold:
- Clearly define and maintain a consistent storage engine abstraction,
  something that is very hard if living inside the kernel where subtle
  abstraction breakages can go unnoticed.
- Allow for breaking out current record-style storage engine implementation
  into its own component as well, only depending on this API, where
  the kernel in the end also would depend on this API and have a much weaker
  dependency on a particular storage engine implementation, through service
  loading or similar.

Even during this breakout some abstractions that was thought to be consistent
and confined wasn't and had to be changed. Such problems immediately surfaced
and was changed accordingly, for example use of cursors and that pooling of
those somehow made it into the interface.

After the introduction of this component, iterative work can begin breaking out
current record-style storage engine into its own component, initially have
kernel depend on that new component. Temporarily disabling that dependency
will highlight places where kernel "cheats" and opportunities to make changes
so that the storage engine API is respected. The dependency can be enabled
and the changes made can be committed and merged into main branch. This way
a potentially massive change set can be broken into smaller change sets.