Removes validation of index updates before commit
In preparation for removing upper limit number of nodes that can exist in
any given index.

This removes a bunch of code related to probing for and reserving space in
an index before committing a transaction and otherwise throwing a specific
IndexCapacityExceededException. All that is now gone and the commit
process and most things around index modifications are simpler.