Fix load of cluster metrics plugin
Previously it was never load since ClusterMembers couldn't be resolved
when loading kernel extensions.  Now we check the edition name (which
has been added to the KernelContext) to verify if we are in HA mode or
not.  Moreover, a Supplier of ClusterMembers is used instead to get
the current instance of ClusterMembers in use in the cluster.

Finally more tests have been added around metrics.