Add the ability of planning index nested loops join
This is not a new operator, but a new way of planning with the existing operators.
The RHS can be planned efficiently by moving data from the LHS, row by row,
and in the RHS this data is used to execute a index seek.

In some situations this is by far the superior alternative to
cartesian products and value hash joins.