Index status check moved into LuceneSchemaIndex
Each schema index in an online state contains a commit with "status"->"online"
map attached as metadata. This metadata is used by LuceneSchemaIndexProvider
to obtain the state of the index (online, populating or failed). IndexingService
uses index state during initialization to create correct index proxy.

Currently index online check lives in the obsolete lucene writer that is knows
nothing about possibly partitioned lucene index and is intended to be removed.

This commit moves the check inside LuceneSchemaIndex and changes
LuceneSchemaIndexProvider to use it instead of the legacy index writer. Commit
with the mentioned user data is stored in the first index partition. Such check
is removed from the legacy writer and the corresponding test.

Verification of index existence is also changed a bit. Now it opens lucene
directory and uses DirectoryReader to check if index exists instead of just
looking at files in the index folder. Such change is needed because in tests we
often use RAMDirectory and file system that are not aware of each other. This
causes index existence check to say that index exists when in fact
RAMDirectory contains some files. Change is included in this commit because
existence check is part of the online check.