Removed IndexStorageFactory#shutdown()
Method was used to close the DirectoryFactory which was passed in as a
constructor dependency. This was not a good idea because DirectoryFactory has
a completely different lifecycle than IndexStorageFactory. Closing of the
in-memory directory factory caused flaky tests because content of in-memory
lucene directories did not survive intentional database restarts.