Fixes a commons lang3 dependency issue
where this component already had that dependency transitively, but the pom
file declared that dependency for use in test scope and so it was no longer
included in runtime scope. This would be problematic if you'd run a main
class from the component directly, but would work fine in a packaged neo4j
since there were other components including that dependency anyway.