Removes accessability of NeoStores from StorageEngine
even through dependency resolution. This forced changes to some places
which had a hard dependency on NeoStores, which was pulled out from
a db or storage engine. Now those places need to access whatever they
did using the kernel API instead and so a couple of methods had to be
added there.

There are still tests which access NeoStores to test some details
which are still valuable to test. These tests can be converted over time
to not depend on NeoStores directly, or to be converted entirely to
tests that revolve around NeoStores. RecordStorageEngine has a method
testAccessNeoStores() for this matter.